silence warning due to kludge.
2006-12-16  Jody Goldberg <jody@gnome.org>

	* ms-excel-read.c (excel_read_COLINFO) : silence warning due to
	  kludge.

2006-12-16  Jody Goldberg <jody@gnome.org>

	Start work on autofilters, only Top10 is implemented for in or out.
	* xlsx-read.c (xlsx_CT_Filters_begin) : new.
	  Massive renaming to use '_begin' rather than '_start' just because I
	  like the sound better.
	* xlsx-write.c (xlsx_write_autofilters) : new.

2006-12-16  Jody Goldberg <jody@gnome.org>

	* src/sheet-filter.h : add GNM_FILTER_OP_BOTTOM_MASK, and
	  GNM_FILTER_OP_PERCENT_MASK.

	* src/xml-sax-read.c (xml_sax_filter_condition) : make bucket count a
	  float.
	* src/xml-sax-write.c (xml_write_filter_field) : ditto.
	* src/xml-io.c (xml_read_filter_field) : ditto.
	* src/sheet-filter.c (gnm_filter_condition_new_bucket) : ditto.
	* gnumeric.xsd : ditto.

+++ src/sheet-filter.c
@@ -65,7 +65,7 @@ gnm_filter_condition_new_double (GnmFilterOp op0, GnmValue *v0,
 }
 
 GnmFilterCondition *
-gnm_filter_condition_new_bucket (gboolean top, gboolean absolute, unsigned n)
+gnm_filter_condition_new_bucket (gboolean top, gboolean absolute, float n)
 {
 	GnmFilterCondition *res = g_new0 (GnmFilterCondition, 1);
 	res->op[0] = GNM_FILTER_OP_TOP_N | (top ? 0 : 1) | (absolute ? 0 : 2);
